made a lot of front end compontnets and addded to db and models

+++ ghi/src/app/signup/page.tsx
@@ -8,14 +8,17 @@ const Page = () => {
   interface FormData {
     username: string;
     password: string;
+    passwordConfirmation: string;
   }
 
-  const [error, setError] = useState<boolean>(false);
+  const [passwordError, setPassworError] = useState<boolean>(false);
+  const [userError, setUserError] = useState<boolean>(false);
   const router = useRouter();
 
   const [formData, setFormData] = useState<FormData>({
     username: "",
     password: "",
+    passwordConfirmation: "",
   });
 
   const handleChange = (
@@ -31,32 +34,39 @@ const Page = () => {
   const handleSubmit = async (e: FormEvent) => {
     e.preventDefault();
 
-    const formDetails = new URLSearchParams();
-    formDetails.append("username", formData.username);
-    formDetails.append("password", formData.password);
+    if (formData.password === formData.passwordConfirmation) {
+      setPassworError(false);
 
-    const loginUrl = "http://localhost:8000/auth/token";
-    const fetchConfig = {
-      method: "post",
-      body: formDetails,
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-    };
-    try {
-      const response = await fetch(loginUrl, fetchConfig);
-      if (response.ok) {
-        const data = await response.json();
-        localStorage.setItem("token", data.access_token);
-        router.push("/characters");
-        console.log("how");
-      } else {
-        setError(true);
-        console.log("Keeps failing");
+      const signUpUrl = "http://localhost:8000/auth/";
+
+      const formDetails = {
+        username: formData.username,
+        password: formData.password,
+      };
+
+      const fetchConfig = {
+        method: "POST",
+        body: JSON.stringify(formDetails),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      };
+      try {
+        const response = await fetch(signUpUrl, fetchConfig);
+        if (response.ok) {
+          setUserError(false);
+          const data = await response.json();
+          console.log(data);
+          localStorage.setItem("token", data.access_token);
+          router.push("/characters");
+        } else {
+          setUserError(true);
+        }
+      } catch (error) {
+        console.log(error);
       }
-    } catch (error) {
-      console.log(error);
-      setError(true);
+    } else {
+      setPassworError(true);
     }
   };
 
@@ -70,9 +80,9 @@ const Page = () => {
             className="mx-auto h-10 w-auto"
           />
           <h2 className="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900 py-4">
-            Sign in to your account
+            Create your Account
           </h2>
-          {error ? (
+          {passwordError ? (
             <div className="rounded-md bg-yellow-50 p-4">
               <div className="flex">
                 <div className="flex-shrink-0">
@@ -86,7 +96,27 @@ const Page = () => {
                     Attention needed
                   </h3>
                   <div className="mt-2 text-sm text-yellow-700">
-                    <p>Your username or password were filled incorrectly</p>
+                    <p>Your passwords do not match</p>
+                  </div>
+                </div>
+              </div>
+            </div>
+          ) : null}
+          {userError ? (
+            <div className="rounded-md bg-yellow-50 p-4">
+              <div className="flex">
+                <div className="flex-shrink-0">
+                  <ExclamationTriangleIcon
+                    aria-hidden="true"
+                    className="h-5 w-5 text-yellow-400"
+                  />
+                </div>
+                <div className="ml-3">
+                  <h3 className="text-sm font-medium text-yellow-800">
+                    Attention needed
+                  </h3>
+                  <div className="mt-2 text-sm text-yellow-700">
+                    <p>Username is already taken</p>
                   </div>
                 </div>
               </div>
@@ -131,13 +161,13 @@ const Page = () => {
                   Password
                 </label>
                 {/* <div className="text-sm">
-                  <a
-                    href="#"
-                    className="font-semibold text-indigo-600 hover:text-indigo-500"
-                  >
-                    Forgot password?
-                  </a>
-                </div> */}
+                      <a
+                        href="#"
+                        className="font-semibold text-indigo-600 hover:text-indigo-500"
+                      >
+                        Forgot password?
+                      </a>
+                    </div> */}
               </div>
               <div className="mt-2">
                 <input
@@ -154,22 +184,53 @@ const Page = () => {
             </div>
 
             <div>
+              <div className="flex items-center justify-between">
+                <label
+                  htmlFor="passwordConfirmation"
+                  className="block text-sm font-medium leading-6 text-gray-900"
+                >
+                  Re-enter your Password
+                </label>
+                {/* <div className="text-sm">
+                      <a
+                        href="#"
+                        className="font-semibold text-indigo-600 hover:text-indigo-500"
+                      >
+                        Forgot password?
+                      </a>
+                    </div> */}
+              </div>
+              <div className="mt-2">
+                <input
+                  id="passwordConfirmation"
+                  name="passwordConfirmation"
+                  type="password"
+                  value={formData.passwordConfirmation}
+                  onChange={handleChange}
+                  required
+                  autoComplete="current-password"
+                  className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                />
+              </div>
+            </div>
+
+            <div>
               <button
                 type="submit"
                 className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
               >
-                Sign in
+                Sign Up
               </button>
             </div>
           </form>
 
           <p className="mt-10 text-center text-sm text-gray-500">
-            Not a member?{" "}
+            Already a user?{" "}
             <a
-              href="/signup"
+              href="/login"
               className="font-semibold leading-6 text-indigo-600 hover:text-indigo-500"
             >
-              Sign up
+              Sign In
             </a>
           </p>
         </div>
